+++ websocket/ws.go
@@ -25,10 +25,9 @@ import (
 // client on rw using the WebSocket subprotocol.
 // It does not perform the WebSocket handshake.
 func NewSession(ctx context.Context, addr jid.JID, rw io.ReadWriter, features ...xmpp.StreamFeature) (*xmpp.Session, error) {
-	n := xmpp.NewNegotiator(func(*xmpp.Session, xmpp.StreamConfig) xmpp.StreamConfig {
+	n := xmpp.NewWebSocketNegotiator(func(*xmpp.Session, xmpp.StreamConfig) xmpp.StreamConfig {
 		return xmpp.StreamConfig{
-			Features:  features,
-			WebSocket: true,
+			Features: features,
 		}
 	})
 	var mask xmpp.SessionState
@@ -42,10 +41,9 @@ func NewSession(ctx context.Context, addr jid.JID, rw io.ReadWriter, features ..
 // receiving server on rw using the WebSocket subprotocol.
 // It does not perform the WebSocket handshake.
 func ReceiveSession(ctx context.Context, rw io.ReadWriter, features ...xmpp.StreamFeature) (*xmpp.Session, error) {
-	n := xmpp.NewNegotiator(func(*xmpp.Session, xmpp.StreamConfig) xmpp.StreamConfig {
+	n := xmpp.NewWebSocketNegotiator(func(*xmpp.Session, xmpp.StreamConfig) xmpp.StreamConfig {
 		return xmpp.StreamConfig{
-			Features:  features,
-			WebSocket: true,
+			Features: features,
 		}
 	})
 	var mask xmpp.SessionState